mbusbed/requirements.txt
aiofiles==24.1.0
aiomqtt @ git+https://github.com/sbtinstruments/asyncio-mqtt@f2647b28387757ef2c7b4a5441b80753ed8c6bce
autopep8==2.3.1
caio==0.9.17
certifi==2024.2.2
charset-normalizer==3.3.2
collection==0.1.6
idna==3.7
nats-py==2.7.2
paho-mqtt==2.1.0
pycodestyle==2.12.0
PyYAML==6.0.1
requests==2.31.0
sdnotify==0.3.2
urllib3==2.2.1
